requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
idna==3.2
pycparser==2.20
pyOpenSSL==20.0.1
requests==2.12.4
six==1.16.0
urllib3==1.26.6
paramiko
